moves frontend decorator into decorators module
